+++ src/components/MapPopup/MapPopup.js
@@ -1,18 +1,35 @@
 import React, { useContext } from 'react';
-import ReactDOMServer from 'react-dom/server';
+import ReactDOM from 'react-dom';
 import Context from '../../DefaultContext';
 import { Box } from '@theme-ui/components';
 import { Popup } from 'mapbox-gl';
 import mapExists from '../../util/mapExists';
+import PopupActionsMenu from './PopupActionsMenu';
 
 const Container = props => {
-  const { styles, properties, config } = props;
+  const { styles, properties, config, showActions, feature } = props;
+
+  let actionsMenu;
+  if (showActions) {
+    if (config.actions && config.actions.length > 0) {
+      actionsMenu = (
+        <PopupActionsMenu popupActions={config.actions} feature={feature} />
+      );
+    } else {
+      console.warn('You passed a showActions prop but no actions');
+      actionsMenu = null;
+    }
+  } else {
+    actionsMenu = null;
+  }
+
   return (
     <Box className="popup-container">
       {config.title && config.title.field ? (
         <Title properties={properties} config={config.title} />
       ) : null}
       <List style={styles} properties={properties} config={config.attributes} />
+      {actionsMenu}
     </Box>
   );
 };
@@ -118,42 +135,52 @@ const List = props => {
 };
 
 const MapPopup = props => {
-  const { layers } = props;
+  const { layers, showActions } = props;
   const config = useContext(Context);
   const map = config.map;
+  const popupContainer = document.createElement('div');
 
   if (!mapExists(map)) {
     return null;
   }
 
   const _build = (config, event) => {
+    const feature = event.features[0];
     if (config.intercept) {
       config
-        .intercept(event.features[0].properties)
+        .intercept(feature.properties)
         .then(function(properties) {
+          ReactDOM.render(
+            React.createElement(Container, {
+              properties,
+              config,
+              showActions,
+              feature: feature ? feature : null
+            }),
+            popupContainer
+          );
+
           new Popup()
             .setLngLat(event.lngLat)
-            .setHTML(
-              ReactDOMServer.renderToString(
-                <Container properties={properties} config={config} />
-              )
-            )
+            .setDOMContent(popupContainer)
             .addTo(map);
         })
         .catch(function(err) {
           console.warn(err);
         });
     } else {
+      ReactDOM.render(
+        React.createElement(Container, {
+          properties: feature.properties,
+          config,
+          showActions,
+          feature: feature ? feature : null
+        }),
+        popupContainer
+      );
       new Popup()
         .setLngLat(event.lngLat)
-        .setHTML(
-          ReactDOMServer.renderToString(
-            <Container
-              properties={event.features[0].properties}
-              config={config}
-            />
-          )
-        )
+        .setDOMContent(popupContainer)
         .addTo(map);
     }
   };